# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sistema de Organización de Eventos — API",
    "version": "1.0.0",
    "description": "API REST del backend del Sistema de Organización de Eventos"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Servidor actual"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token"
      }
    },
    "schemas": {
...
      "PlanPago": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "id_evento": {
            "type": "integer"
          },
          "numero_cuota": {
            "type": "integer"
          },
          "monto": {
            "type": "string"
          },
          "fecha_limite": {
            "type": "string",
            "format": "date"
          },
          "pagada": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PagoCliente": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "id_evento": {
            "type": "integer"
          },
          "id_cuota": {
            "type": "integer",
            "nullable": true
          },
          "monto": {
            "type": "string"
          },
          "fecha_pago": {
            "type": "string",
            "format": "date"
          },
          "metodo_pago": {
            "type": "string",
            "enum": [
              "Efectivo",
              "Transferencia",
              "Tarjeta",
              "QR"
            ]
          },
          "comprobante_url": {
            "type": "string",
            "nullable": true
          },
          "registrado_por": {
            "type": "string",
            "enum": [
              "Cliente",
              "Administrador"
            ]
          },
          "estado": {
            "type": "string",
            "enum": [
              "Reportado",
              "Validado",
              "Rechazado"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PagoProveedor": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "id_evento_proveedor": {
            "type": "integer"
          },
          "monto": {
            "type": "string"
          },
          "fecha_pago": {
            "type": "string",
            "format": "date"
          },
          "metodo_pago": {
            "type": "string",
            "enum": [
              "Efectivo",
              "Transferencia",
              "Tarjeta",
              "QR"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/eventos/{id}/plan-pagos": {
      "get": {
        "summary": "Consulta el plan de cuotas del evento (incluye campo derivado `pagada` por cuota)",
        "tags": [
          "Pagos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cuotas del plan",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlanPago"
                  }
                }
              }
            }
          },
          "403": {
            "description": "No autorizado (solo Administrador)"
          },
          "404": {
            "description": "Evento no encontrado"
          }
        }
      },
      "post": {
        "summary": "Define el plan de cuotas del evento (solo si el evento no tiene uno todavía)",
        "tags": [
          "Pagos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "cuotas"
                ],
                "properties": {
                  "cuotas": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "numero_cuota",
                        "monto",
                        "fecha_limite"
                      ],
                      "properties": {
                        "numero_cuota": {
                          "type": "integer"
                        },
                        "monto": {
                          "type": "number"
                        },
                        "fecha_limite": {
                          "type": "string",
                          "format": "date"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Plan creado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlanPago"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Body inválido"
          },
          "403": {
            "description": "No autorizado (solo Administrador)"
          },
          "409": {
            "description": "El evento ya tiene un plan de pagos definido"
          }
        }
      }
    },
    "/plan-pagos/{cuotaId}": {
      "put": {
        "summary": "Actualiza una cuota del plan",
        "tags": [
          "Pagos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "cuotaId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "numero_cuota": {
                    "type": "integer"
                  },
                  "monto": {
                    "type": "number"
                  },
                  "fecha_limite": {
                    "type": "string",
                    "format": "date"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cuota actualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanPago"
                }
              }
            }
          },
          "403": {
            "description": "No autorizado (solo Administrador)"
          },
          "404": {
            "description": "Cuota no encontrada"
          }
        }
      }
    },
    "/eventos/{id}/pagos": {
      "get": {
        "summary": "Lista los pagos (ingresos) del evento",
        "tags": [
          "Pagos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de pagos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PagoCliente"
                  }
                }
              }
            }
          },
          "403": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Evento no encontrado"
          }
        }
      },
      "post": {
        "summary": "Registra un pago del Cliente. Si lo registra el Cliente queda en estado Reportado (pendiente de validación); si lo registra el Administrador queda Validado de inmediato.\n",
        "tags": [
          "Pagos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "monto",
                  "fecha_pago",
                  "metodo_pago"
                ],
                "properties": {
                  "id_cuota": {
                    "type": "integer"
                  },
                  "monto": {
                    "type": "number"
                  },
                  "fecha_pago": {
                    "type": "string",
                    "format": "date"
                  },
                  "metodo_pago": {
                    "type": "string",
                    "enum": [
                      "Efectivo",
                      "Transferencia",
                      "Tarjeta",
                      "QR"
                    ]
                  },
                  "comprobante": {
                    "type": "string",
                    "format": "binary",
                    "description": "Opcional",
                    "PDF/JPG/PNG": null
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pago registrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagoCliente"
                }
              }
            }
          },
          "400": {
            "description": "Faltan campos obligatorios",
            "o la cuota no pertenece a este evento": null
          },
          "403": {
            "description": "No autorizado"
          }
        }
      }
    },
    "/pagos/{id}/validar": {
      "put": {
        "summary": "Valida o rechaza un pago reportado por el Cliente",
        "tags": [
          "Pagos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "decision"
                ],
                "properties": {
                  "decision": {
                    "type": "string",
                    "enum": [
                      "Validado",
                      "Rechazado"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pago actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagoCliente"
                }
              }
            }
          },
          "400": {
            "description": "decision inválida, o el pago no está en estado Reportado"
          },
          "403": {
            "description": "No autorizado (solo Administrador)"
          },
          "404": {
            "description": "Pago no encontrado"
          }
        }
      }
    },
    "/eventos/{id}/egresos": {
      "get": {
        "summary": "Consulta los pagos (egresos) a proveedores del evento",
        "tags": [
          "Pagos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de egresos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PagoProveedor"
                  }
                }
              }
            }
          },
          "403": {
            "description": "No autorizado (solo Administrador)"
          },
          "404": {
            "description": "Evento no encontrado"
          }
        }
      },
      "post": {
        "summary": "Registra un pago a un proveedor contratado. Al registrarlo se recalcula automáticamente el estado_contrato de la contratación a Pagado si el total pagado alcanza el costo_acordado.\n",
        "tags": [
          "Pagos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id_evento_proveedor",
                  "monto",
                  "fecha_pago",
                  "metodo_pago"
                ],
                "properties": {
                  "id_evento_proveedor": {
                    "type": "integer"
                  },
                  "monto": {
                    "type": "number"
                  },
                  "fecha_pago": {
                    "type": "string",
                    "format": "date"
                  },
                  "metodo_pago": {
                    "type": "string",
                    "enum": [
                      "Efectivo",
                      "Transferencia",
                      "Tarjeta",
                      "QR"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Egreso registrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagoProveedor"
                }
              }
            }
          },
          "400": {
            "description": "Faltan campos obligatorios",
            "o la contratación no pertenece a este evento": null
          },
          "403": {
            "description": "No autorizado (solo Administrador)"
          }
        }
      }
    },
...
  },
...
}